Truncate log instead of deleting
This is unfortunately also not great for crash safety: a crash in
between installing the new table and truncating the log will result in
adding a duplicate of the table. This is technically correct, but it
would be better to log the log truncate command in the manifest and then
mark it done; on crash, the manifest recovery can see that the log's
table was installed but the log wasn't truncated and thus truncate the
log without adding a new table.