Pre-encode log transaction before writing
Since the log file is unbuffered, this causes the entire log write to be
a single system call, which is vastly more efficient (previously we
needed four system calls, three of which were 1-2 bytes). This could
instead be implemented generically using bufio, but this is probably a
simpler solution even in Go (due to the complexity of adding flushes in
the right places with that solution).